+++ Editor/ARCoreRuntimeSettingsEditor.cs
@@ -1,3 +1,6 @@
+#if UNITY_6000_5_OR_NEWER
+using Unity.Scripting.LifecycleManagement;
+#endif
 using UnityEngine;
 using UnityEngine.XR.ARCore;
 
@@ -20,7 +23,14 @@ namespace UnityEditor.XR.ARCore
         static readonly GUIContent k_KeylessAuthUrlText = new("View Documentation");
         const string k_KeylessAuthUrl = "https://developers.google.com/ar/develop/authorization";
 
+#if UNITY_6000_5_OR_NEWER
+        [NoAutoStaticsCleanup]
+#endif
         static GUIStyle s_UrlLabelPersonal;
+
+#if UNITY_6000_5_OR_NEWER
+        [NoAutoStaticsCleanup]
+#endif
         static GUIStyle s_UrlLabelProfessional;
 
         void OnEnable()